arduino: remove wifi conn responsibility from comm
Before, the responsibility of connecting to wifi is in comm_wifi's.
This seems like bad design because it creates side affects within the
constructor, and seems coupled. Instead, provide a begin() function
(defaulted to a noop) in comm's interface, which performs the wifi
connection. It is the main file's responsibility to call this function.